Preentrega-3-cambios-alertas

+++ js/carrito.js
@@ -11,26 +11,26 @@ let botonesEliminar = document.querySelectorAll('.producto__eliminar');
 loginUser = JSON.parse(localStorage.getItem('login-user'));
 
 
-function CargarProductosCarrito(){
-    if (!ProductosEnCarrito || !loginUser || ProductosEnCarrito.length === 0  ){
+function CargarProductosCarrito() {
+    if (!ProductosEnCarrito || !loginUser || ProductosEnCarrito.length === 0) {
         window.location.href = '../index.html'
-    }else {
-    
+    } else {
+
         ContenedorPorducto.innerHTML = '';
-    
+
         ProductosEnCarrito.forEach(producto => {
             const plantilla = plantillaProducto.cloneNode(true);
-    
+
             plantilla.querySelector('.producto__img').src = `.${producto.src}`;
             plantilla.querySelector('.titulo').innerText = producto.nombre;
             plantilla.querySelector('.cantidad').innerText = producto.cantidad;
             plantilla.querySelector('.precio').innerText = producto.precio;
             plantilla.querySelector('.subtotal').innerText = `$${producto.precio * producto.cantidad}`;
             plantilla.querySelector('.producto__eliminar').id = producto.id;
-            
+
             ContenedorPorducto.appendChild(plantilla);
         });
-    
+
     }
 
     RecuperarBotonesEliminar()
@@ -41,7 +41,7 @@ CargarProductosCarrito()
 
 
 
-function RecuperarBotonesEliminar(){
+function RecuperarBotonesEliminar() {
     botonesEliminar = document.querySelectorAll('.producto__eliminar');
 
     botonesEliminar.forEach(boton => {
@@ -49,7 +49,23 @@ function RecuperarBotonesEliminar(){
     });
 }
 
-function eliminarDelCarrito(e){
+function eliminarDelCarrito(e) {
+    Toastify({
+
+        text: "Se elimino del carrito",
+        duration: 1500,
+        close: false,
+        gravity: "top", // `top` or `bottom`
+        position: "left", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+            background: "linear-gradient(to right, #2D1A40, #2D1A40)",
+            borderRadius: '2rem',
+            textTransform: 'uppercase',
+            fontSize: '.8rem'
+        },
+        onClick: function () { } // Callback after click
+    }).showToast();
     const idBoton = e.currentTarget.id;
     const index = ProductosEnCarrito.findIndex(producto => producto.id === idBoton);
     ProductosEnCarrito.splice(index, 1);
@@ -59,19 +75,34 @@ function eliminarDelCarrito(e){
     localStorage.setItem('productos', JSON.stringify(ProductosEnCarrito));
 }
 
-function VaciarCarrito(){
-    ProductosEnCarrito.length = 0;
-    localStorage.setItem('productos', JSON.stringify(ProductosEnCarrito));
-    CargarProductosCarrito()
+function VaciarCarrito() {
+    Swal.fire({
+        title: "Estas Seguro?",
+        icon: "question",
+        html: `Se van a borrar ${ProductosEnCarrito.reduce((acc, producto)=> acc + producto.cantidad,0)} productos`,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: `Si`,
+        cancelButtonText: `No`,
+        cancelButtonAriaLabel: "Thumbs down"
+    }).then((result) => {
+        /* Read more about isConfirmed, isDenied below */
+        if (result.isConfirmed) {
+            ProductosEnCarrito.length = 0;
+            localStorage.setItem('productos', JSON.stringify(ProductosEnCarrito));
+            CargarProductosCarrito()
+        }
+    });
+
 }
 
 
-function ActualizarTotal(){
-    const Total = ProductosEnCarrito.reduce((acc, producto) => acc + (producto.precio * producto.cantidad),0);
+function ActualizarTotal() {
+    const Total = ProductosEnCarrito.reduce((acc, producto) => acc + (producto.precio * producto.cantidad), 0);
     carritoTotal.innerText = `$${Total}`;
 }
 
-function ComprarCarrito(){
+function ComprarCarrito() {
     Swal.fire({
         position: "center",
         icon: "success",
@@ -79,12 +110,12 @@ function ComprarCarrito(){
         showConfirmButton: false,
         timer: 1500
     });
-    setTimeout(()=>{
+    setTimeout(() => {
         ProductosEnCarrito.length = 0;
         localStorage.setItem('productos', JSON.stringify(ProductosEnCarrito));
         window.location.href = '../index.html'
-    },1500)
-    
+    }, 1500)
+
 }
 
 CarritoVacio.addEventListener('click', VaciarCarrito)
